+++ common/cpp/react/renderer/components/TrueSheetSpec/TrueSheetContentViewState.cpp
@@ -4,7 +4,8 @@ namespace facebook::react {
 
 #ifdef ANDROID
 folly::dynamic TrueSheetContentViewState::getDynamic() const {
-  return folly::dynamic::object("scrollableBounded", scrollableBounded);
+  return folly::dynamic::object("scrollableBounded", scrollableBounded)(
+      "naturalHeight", naturalHeight);
 }
 #endif